revision_local_app/requirements.txt
streamlit>=1.43.0
pypdf>=5.4.0
python-docx>=1.1.2
python-pptx>=1.0.2
opencv-python-headless>=4.11.0.86
pytesseract>=0.3.13
Pillow>=11.1.0
reportlab>=4.3.1
numpy>=2.2.4
psutil>=7.2.2
